function/requirements.txt
azure-functions==1.18.0
azure-storage-blob==12.19.0
azure-ai-vision-imageanalysis==1.0.0
azure-identity==1.15.0
applicationinsights==0.11.10
